@@ -1,53 +1,55 @@
 {
-  "name": "uidhtml-modular-website",
-  "version": "0.0.0",
-  "scripts": {
-    "ng": "ng",
-    "start": "ng serve",
-    "build": "ng build",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
-  },
-  "private": true,
-  "dependencies": {
-    "@angular/animations": "^6.1.0",
-    "@angular/cdk": "^6.4.3",
-    "@angular/common": "^6.1.0",
-    "@angular/compiler": "^6.1.0",
-    "@angular/core": "^6.1.0",
-    "@angular/forms": "^6.1.0",
-    "@angular/http": "^6.1.0",
-    "@angular/material": "^6.4.3",
-    "@angular/platform-browser": "^6.1.0",
-    "@angular/platform-browser-dynamic": "^6.1.0",
-    "@angular/router": "^6.1.0",
-    "core-js": "^2.5.4",
-    "hammerjs": "^2.0.8",
-    "rxjs": "^6.0.0",
-    "zone.js": "~0.8.26"
-  },
-  "devDependencies": {
-    "@angular-devkit/build-angular": "~0.7.0",
-    "@angular/cli": "~6.1.1",
-    "@angular/compiler-cli": "^6.1.0",
-    "@angular/language-service": "^6.1.0",
-    "@types/jasmine": "~2.8.6",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.2.1",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~1.7.1",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.0",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "ng2-ckeditor": "^1.2.1",
-    "ngx-youtube-player": "^6.0.0",
-    "protractor": "^5.4.0",
-    "ts-node": "~5.0.1",
-    "tslint": "~5.9.1",
-    "typescript": "~2.7.2"
-  }
+    "name": "uidhtml-modular-website",
+    "version": "0.0.0",
+    "scripts": {
+        "ng": "ng",
+        "start": "ng serve --open",
+        "build": "ng build",
+        "test": "ng test",
+        "lint": "ng lint",
+        "e2e": "ng e2e"
+    },
+    "private": true,
+    "dependencies": {
+        "@angular/animations": "^7.1.1",
+        "@angular/cdk": "^7.1.0",
+        "@angular/common": "^7.1.1",
+        "@angular/compiler": "^7.1.1",
+        "@angular/core": "^7.1.1",
+        "@angular/forms": "^7.1.1",
+        "@angular/http": "^7.1.1",
+        "@angular/material": "^7.1.0",
+        "@angular/platform-browser": "^7.1.1",
+        "@angular/platform-browser-dynamic": "^7.1.1",
+        "@angular/router": "^7.1.1",
+        "core-js": "*",
+        "hammerjs": "^2.0.8",
+        "highlight.js": "^9.13.1",
+        "ng2-ckeditor": "^1.2.2",
+        "ngx-highlightjs": "^3.0.0",
+        "rxjs": "^6.3.3",
+        "zone.js": "^0.8.26"
+    },
+    "devDependencies": {
+        "@angular-devkit/build-angular": "^0.11.0",
+        "@angular/cli": "^7.1.0",
+        "@angular/compiler-cli": "^7.1.1",
+        "@angular/language-service": "^7.1.1",
+        "@types/jasmine": "^3.3.0",
+        "@types/jasminewd2": "^2.0.6",
+        "@types/node": "^10.12.11",
+        "codelyzer": "^4.5.0",
+        "jasmine-core": "^3.3.0",
+        "jasmine-spec-reporter": "*",
+        "karma": "^3.1.2",
+        "karma-chrome-launcher": "*",
+        "karma-coverage-istanbul-reporter": "^2.0.4",
+        "karma-jasmine": "^2.0.1",
+        "karma-jasmine-html-reporter": "^1.4.0",
+        "ngx-youtube-player": "*",
+        "protractor": "^5.4.1",
+        "ts-node": "^7.0.1",
+        "tslint": "^5.11.0",
+        "typescript": "^3.1.1"
+    }
 }
