@@ -1,4 +1,4 @@
-import { Component, OnInit, Output, EventEmitter } from '@angular/core';
+import { Component, OnInit, Output, Input, EventEmitter } from '@angular/core';
 
 @Component({
   selector: 'app-ck-editor',
@@ -7,19 +7,45 @@ import { Component, OnInit, Output, EventEmitter } from '@angular/core';
 })
 export class CkEditorComponent implements OnInit {
   public ckeConfig: any;
+  @Input() ckeditorContent: string = '';
   @Output() public bodyContent = new EventEmitter<any>();
-  constructor() { }
+  constructor() {
+   }
 
   ngOnInit() {
     this.ckeConfig = {
-      allowedContent: true,
-      extraPlugins: 'divarea',
-      forcePasteAsPlainText: false
+      uiColor: '#F0F3F4'
     };
   }
   onChange($event: any): void {
-    console.log($event);
+    // console.log(this.ckeditorContent);
     this.bodyContent.emit($event);
-    // this.log += new Date() + "<br />";
+  }
+  onEditorChange($event: any): void {
+    // console.log($event);
+  }
+  onReady($event: any): void {
+    // console.log($event);
+  }
+  onFocus($event: any): void {
+    // console.log($event);
+  }
+  onBlur($event: any): void {
+    // console.log($event);
+  }
+  onContentDom($event: any): void {
+    // console.log($event);
+  }
+  onFileUploadRequest($event: any): void {
+    // console.log($event);
+  }
+  onFileUploadResponse($event: any): void {
+    // console.log($event);
+  }
+  onPaste($event: any): void {
+    // console.log($event);
+  }
+  onDrop($event: any): void {
+    // console.log($event);
   }
 }
