@@ -17,6 +17,8 @@ export class PostFormComponent implements OnInit {
   public zipError: boolean = false;
   public createdDate: string = '';
   public total_record_in_table: number = 0;
+  public defaultPreviewImage: string = '../../../../../../../assets/images/upload-image-icon.png';
+  public modalError: string = '';
 
   public form: FormGroup;
 
@@ -41,18 +43,27 @@ export class PostFormComponent implements OnInit {
       metaDesc: ['', Validators.required],
       keywords: ['', Validators.required],
       body: ['', Validators.required],
+      htmlCode: ['hello'],
+      cssCode: [''],
+      javascriptCode: [''],
+      angularCode: [''],
+      jqueryCode: [''],
+      typescriptCode: [''],
+      phpCode: [''],
       status: 0
     });
   }
 
   ngOnInit() {
+    this.getDataFromDB();
+  }
+  getDataFromDB() {
     const tableName: string = 'posts';
     const url: string = '/src/app/_apis/admin/get_total_record_from_table.php?tableName=' + tableName;
     this._submitFormService._getData(url).subscribe((data) => {
       this.total_record_in_table = data[0].row_num + 1;
     });
   }
-
   // Traget file type input and apply click event to get image or zip file
   browseFile(event): void {
     event.preventDefault();
@@ -67,8 +78,8 @@ export class PostFormComponent implements OnInit {
     if ($event.currentTarget.id === 'imageFileField') {
       this.formData.append('imageFile', file);
       this.imageError = this.checkImageError('image');
-      if (file.size >= 102400) {
-        this.imageErrorData = ' File size should be less then 100KB ';
+      if (file.size >= 1002400) {
+        this.imageErrorData = ' File size should be less then 1MB ';
         this.form.controls['image'].setValue('');
         this.imageError = this.checkImageError('image');
       } else {
@@ -94,6 +105,24 @@ export class PostFormComponent implements OnInit {
     this.form.controls['body'].setValue($event);
   }
 
+  getExampleCode(event) {
+    if (event.target.name === 'htmlCode') {
+      this.form.controls['htmlCode'].setValue(event.target.value);
+    } else if (event.target.name === 'cssCode') {
+      this.form.controls['cssCode'].setValue(event.target.value);
+    } else if (event.target.name === 'javascriptCode') {
+      this.form.controls['javascriptCode'].setValue(event.target.value);
+    } else if (event.target.name === 'angularCode') {
+      this.form.controls['angularCode'].setValue(event.target.value);
+    } else if (event.target.name === 'jqueryCode') {
+      this.form.controls['jqueryCode'].setValue(event.target.value);
+    } else if (event.target.name === 'typescriptCode') {
+      this.form.controls['typescriptCode'].setValue(event.target.value);
+    } else if (event.target.name === 'typescriptCode') {
+      this.form.controls['typescriptCode'].setValue(event.target.value);
+    }
+  }
+
   // Check for image and zip error
   checkImageError(controlName: string): boolean {
     let error = false;
@@ -106,20 +135,40 @@ export class PostFormComponent implements OnInit {
     return error;
   }
   // Submit form
-  submit() {
+  onSubmit(event) {
     this.imageError = this.checkImageError('image');
     Object.keys(this.form.value).forEach((key) => this.formData.append(key, this.form.value[key]));
 
     const url: string = '/src/app/_apis/admin/add-post.php';
     this._submitFormService._postData(url, this.formData).subscribe((data) => {
-      this.openDialog();
-      console.log(data);
+      this.openDialog(event, data[0].msg);
+      if (data[0].result === 1) {
+        this.resetForm();
+        this.previewImage.nativeElement.src = this.defaultPreviewImage;
+        this.showImageName.nativeElement.value = '';
+        this.showZipName.nativeElement.value = '';
+        this.form.controls['status'].setValue(0);
+        this.form.controls['createdDate'].setValue(new Date());
+        this.form.controls['body'].setValue('');
+      }
     });
+    event.preventDefault();
+    event.stopPropagation();
   }
   // Open dialog for result
-  openDialog(): void {
+  openDialog(event, msg): void {
     const dialogRef = this.dialog.open(MaterialDialogueComponent, {
       width: '250px',
-      data: { header: 'Upload Report', content: 'uploaded' }
+      data: { header: 'Post status', content: msg }
     });
+    dialogRef.afterClosed().subscribe(result => {
+      event.preventDefault();
+      event.stopPropagation();
+      this.getDataFromDB();
+    });
+  }
+
+  resetForm() {
+    this.form.reset();
+  }
 }
